backend/requirements.txt
fastapi
uvicorn
requests
httpx
pinecone-client
python-dotenv
langchain
langgraph
langchain-community
langchain-cohere
cohere